requirements.txt
alembic==1.16.1
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.40.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pqcrypto==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.23.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.2.3
pypqc==0.0.6.2
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
semantic-version==2.10.0
setuptools-rust==1.11.1
simple-websocket==1.1.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
